// 电影推荐系统前端源码/film-recommand-sys/src/api/film/api.ts
export enum Api {
  getFilmList = "film/list",
  updateFilm = "film/update",
  getRecommendFilm = "film/recommendFilm",
  getFilmInfo = "film/getFilmInfo/",
  getAppraiseList = "appraise/list",
  saveAppraise = "appraise/save",
  delAppraise = "appraise/del/",
  getThumbList = "thumb/list",
  saveThumb = "thumb/save",
  updateThumb = "thumb/update",
  delThumb = "thumb/del/",
  getThumbNum = "thumb/thumbNum/",

  getFilmTop10 = "film/top",
}